Moved globalSettings handling into base
